+++ 0x08-recursion/main.h
@@ -5,6 +5,6 @@ int _strlen_recursion(char *s);
 int factorial(int n);
 int _pow_recursion(int x, int y);
 int _sqrt_recursion(int n);
-int squareroot(int n, int i);
+int squareroot(int n, long i);
 int is_prime_number(int n);
 int prime_helper(long i, int n);
